% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_to_arena.R
\name{excel_to_arena}
\alias{excel_to_arena}
\title{Import data to Arena compatible txt file}
\usage{
excel_to_arena(data)
}
\arguments{
\item{vector}{}
}
\value{
txt file
}
\description{
Import data to Arena compatible txt file
}
